+++ lib/karafka/base_responder.rb
@@ -198,7 +198,9 @@ module Karafka
     # @param options [Hash] options for waterdrop
     # @return [Class] WaterDrop producer (sync or async based on the settings)
     def producer(options)
-      options[:async] ? WaterDrop::AsyncProducer : WaterDrop::SyncProducer
+      self.class.topics[
+        options[:topic]
+      ].async? ? WaterDrop::AsyncProducer : WaterDrop::SyncProducer
     end
   end
 end