+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ## 1.2.5
 - #354 - Expose consumer heartbeat
+- #373 - Async producer not working properly with responders
 
 ## 1.2.4
 - #332 - Fetcher for max queue size